+++ js/main.js
@@ -1,3 +1,4 @@
 import './rendering.js';
 import './big-photo.js';
 import './img-upload.js';
+import './validation.js';
